+++ src/moduler/aktivitet/visning/avtalt-container/ForhaandsorienteringBrukerVisning.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import { useDispatch } from 'react-redux';
 
 import { Aktivitet } from '../../../../datatypes/aktivitetTypes';
-import { markerForhaandsorienteringSomLest } from '../../aktivitet-actions';
-import { markerForhaandsorienteringSomLestArenaAktivitet } from '../../arena-aktiviteter-reducer';
 import DeleLinje from '../delelinje/delelinje';
 import Forhaandsorienteringsvisning from '../hjelpekomponenter/forhaandsorientering/Forhaandsorienteringsvisning';
 
@@ -11,29 +8,12 @@ interface Props {
     aktivitet: Aktivitet;
 }
 
-const ForhaandsorienteringBrukerVisning = (props: Props) => {
-    const { aktivitet } = props;
-
-    const dispatch = useDispatch();
-
-    const onMarkerSomLest = () => {
-        if (aktivitet.arenaAktivitet) {
-            dispatch(markerForhaandsorienteringSomLestArenaAktivitet(aktivitet));
-        } else {
-            dispatch(markerForhaandsorienteringSomLest(aktivitet));
-        }
-    };
-
-    return (
-        <>
-            <DeleLinje />
-            <Forhaandsorienteringsvisning
-                forhaandsorientering={aktivitet.forhaandsorientering}
-                onMarkerSomLest={onMarkerSomLest}
-            />
-            <DeleLinje />
-        </>
-    );
-};
+const ForhaandsorienteringBrukerVisning = ({ aktivitet }: Props) => (
+    <>
+        <DeleLinje />
+        <Forhaandsorienteringsvisning aktivitet={aktivitet} />
+        <DeleLinje />
+    </>
+);
 
 export default ForhaandsorienteringBrukerVisning;
